fastapi/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
slowapi==0.1.9

# Database
mysql-connector-python==9.3.0
python-dotenv==1.1.0

# Web Automation
selenium==4.33.0

# Background Tasks (Future)
celery==5.4.0
redis==5.2.1

# Utilities
pillow==11.2.1
pandas==2.3.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-timeout==2.1.0
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0